Part: introduce RefinableFeature base for boolean refine properties
Add Part::RefinableFeature (inherits Part::Feature) as a shared base
for Boolean, MultiFuse, and MultiCommon. It owns the Refine and
CheckRefine properties and the logic to validate and apply the refine
step, which was previously duplicated across all three classes.

- RefinableFeature::isRefineResultValid() — BOPAlgo self-intersection
  check, available to subclasses (used directly by MultiFuse which has
  its own history-tracking refine path)
- RefinableFeature::applyRefine() — applies makeElementRefine(), reverts
  and warns if the result is invalid; used by Boolean and MultiCommon
- Preference helpers and BOPAlgo/BRepBuilderAPI_Copy includes move to
  FeaturePartRefinable.cpp
- Boolean::execute() and MultiCommon::execute() refine blocks collapse
  to a single this->applyRefine(res) call

Requested in review of PR #29134.
