+++ src/Mod/Part/App/FeaturePartFuse.cpp
@@ -45,9 +45,6 @@ using namespace Part;
 namespace Part
 {
 extern void throwIfInvalidIfCheckModel(const TopoDS_Shape& shape);
-extern bool getRefineModelParameter();
-extern bool getCheckRefineParameter();
-extern bool refineResultIsValid(const TopoDS_Shape& shape, bool checkRefine);
 }  // namespace Part
 
 PROPERTY_SOURCE(Part::Fuse, Part::Boolean)
@@ -68,7 +65,7 @@ const char* Fuse::opCode() const
 
 // ----------------------------------------------------
 
-PROPERTY_SOURCE(Part::MultiFuse, Part::Feature)
+PROPERTY_SOURCE(Part::MultiFuse, Part::RefinableFeature)
 
 
 MultiFuse::MultiFuse()
@@ -84,23 +81,6 @@ MultiFuse::MultiFuse()
     );
     History.setSize(0);
 
-    ADD_PROPERTY_TYPE(
-        Refine,
-        (0),
-        "Boolean",
-        (App::PropertyType)(App::Prop_None),
-        "Refine shape (clean up redundant edges) after this boolean operation"
-    );
-    ADD_PROPERTY_TYPE(
-        CheckRefine,
-        (false),
-        "Boolean",
-        (App::PropertyType)(App::Prop_None),
-        "Validate refine result and revert if it introduces self-intersections (slower)"
-    );
-
-    this->Refine.setValue(getRefineModelParameter());
-    this->CheckRefine.setValue(getCheckRefineParameter());
 }
 
 short MultiFuse::mustExecute() const
@@ -192,7 +172,7 @@ App::DocumentObjectExecReturn* MultiFuse::execute()
                 try {
                     TopoDS_Shape oldShape = res.getShape();
                     BRepBuilderAPI_RefineModel mkRefine(oldShape);
-                    if (!refineResultIsValid(mkRefine.Shape(), this->CheckRefine.getValue())) {
+                    if (!this->isRefineResultValid(mkRefine.Shape())) {
                         Base::Console().warning(
                             "'%s': The boolean result is correct, but the "
                             "Refine (cleanup) step damaged it and was skipped. "
