+++ src/Mod/Part/App/FeaturePartFuse.h
@@ -48,7 +48,7 @@ protected:
     //@}
 };
 
-class PartExport MultiFuse: public Part::Feature
+class PartExport MultiFuse: public Part::RefinableFeature
 {
     PROPERTY_HEADER_WITH_OVERRIDE(Part::MultiFuse);
 
@@ -57,8 +57,6 @@ public:
 
     App::PropertyLinkList Shapes;
     PropertyShapeHistory History;
-    App::PropertyBool Refine;
-    App::PropertyBool CheckRefine;
 
     /** @name methods override feature */
     //@{
